capstone/airesume/requirements.txt
Django
python-dotenv
xhtml2pdf
openai